Move public mocked methods into a trait
