+++ tests/Mock/MockGitTool.php
@@ -4,11 +4,14 @@ declare(strict_types = 1);
 namespace MountHolyoke\JorgeTests\Mock;
 
 use MountHolyoke\Jorge\Tool\GitTool;
+use MountHolyoke\JorgeTests\Mock\MockToolPublicMethodsTrait;
 
 /**
  * Supplants the GitTool class so we can test things.
  */
 class MockGitTool extends GitTool {
+  use MockToolPublicMethodsTrait;
+
   /**
    * {@inheritDoc}
    */
